nvi: fix ex_equal %ld/recno_t varargs mismatch
lno is recno_t (uint32_t) but was passed to %ld; ex_printf forwards
the va_list to vsnprintf, so this is a real size/signedness mismatch
(UB, benign on LP64 little-endian, wrong on big-endian). Use nvi's
house idiom %lu/(u_long). Verbatim upstream.

+++ src.freebsd/nvi/ex/ex_equal.c
@@ -49,6 +49,6 @@ ex_equal(SCR *sp, EXCMD *cmdp)
 	} else
 		lno = cmdp->addr1.lno;
 
-	(void)ex_printf(sp, "%ld\n", lno);
+	(void)ex_printf(sp, "%lu\n", (u_long)lno);
 	return (0);
 }
